massive restructure and renaming for hopefully more transparent coding
