+++ connectomix/core/processing/stats.py
@@ -1,5 +1,6 @@
 import numpy as np
 from nilearn.connectome import sym_matrix_to_vec, vec_to_sym_matrix
+from nilearn.glm.second_level import non_parametric_inference
 from scipy.stats import permutation_test
 from statsmodels.regression.linear_model import OLS
 from statsmodels.tools import add_constant
@@ -21,7 +22,7 @@ def generate_permuted_null_distributions(group_data, config, layout, entities, o
                                                        extension=".npy",
                                                        suffix="permutations",
                                                        return_type="filename")
-    from connectomix.utils.tools import apply_nonbids_filter
+    from connectomix.core.utils.bids import apply_nonbids_filter
     perm_files = apply_nonbids_filter("analysis",
                                       config["analysis_label"],
                                       perm_files)
@@ -45,8 +46,8 @@ def generate_permuted_null_distributions(group_data, config, layout, entities, o
                                                                  path_patterns=[
                                                                      "group/{analysis_label}/permutations/group_[ses-{session}_][run-{run}_][task-{task}]_space-{space}_method-{method}_desc-{desc}_analysis-{analysis_label}_permutations.npy"],
                                                                  validate=False)
-        from connectomix.utils.makers import ensure_directory
-        ensure_directory(perm_file)
+        from connectomix.core.utils.tools import make_parent_dir
+        make_parent_dir(perm_file)
         # If nothing has to be loaded, then initiate the null distribution with the observed values
         perm_data = np.array([list(observed_stats.values())])  # Size is (1,2) and order is max followed by min
 
@@ -80,7 +81,6 @@ def generate_permuted_null_distributions(group_data, config, layout, entities, o
     print(".")
     print("Permutations computed.")
     return perm_data.reshape([-1, 1])[0:], perm_data.reshape([-1, 1])[1:]  # Returning all maxima and all minima
-
 
 # Regression analysis of each connectivity value with a covariate, with optionnal confounds and optional permuted columns
 def regression_analysis(group_data, design_matrix, permutate=False):
@@ -143,7 +143,6 @@ def regression_analysis(group_data, design_matrix, permutate=False):
 
     return t_values_matrix, p_values_matrix
 
-
 # Define a function to compute the difference in connectivity between the two groups
 # Todo: adapt this for paired tests
 def stat_func(x, y):
@@ -167,3 +166,45 @@ def stat_func(x, y):
     t_stat, _ = ttest_ind(x, y)
     return t_stat
 
+
+def compute_non_parametric_max_mass(layout, glm, label, config):
+    from connectomix.core.utils.loaders import load_entities_from_config
+    entities = load_entities_from_config(config)
+    entities.pop("subject")
+
+    from connectomix.core.utils.bids import build_output_path
+    np_logp_max_mass_path = build_output_path(layout,
+                                              entities,
+                                              label,
+                                              "group",
+                                              config,
+                                              suffix="logpMaxMass",
+                                              extension=".nii.gz")
+
+    # entities["seed"] = label
+    #
+    # np_logp_max_mass_path = layout.derivatives["connectomix"].build_path({**entities,
+    #                                                                       "analysis_label": config["analysis_label"],
+    #                                                                       "method": config["method"],
+    #                                                                       "seed": label
+    #                                                                       },
+    #                                                                      path_patterns=[
+    #                                                                          "group/{analysis_label}/group_[ses-{session}_][run-{run}_][task-{task}]_space-{space}_method-{method}_seed-{seed}_analysis-{analysis_label}_logpMaxMass.nii.gz"],
+    #                                                                      validate=False)
+    #
+    # from connectomix.core.makers import ensure_directory
+    # ensure_directory(np_logp_max_mass_path)
+
+    np_outputs = non_parametric_inference(glm.second_level_input_,
+                                          design_matrix=glm.design_matrix_,
+                                          second_level_contrast=config["contrast"],
+                                          first_level_contrast=label,
+                                          smoothing_fwhm=config["smoothing"],
+                                          two_sided_test=config["two_sided_test"],
+                                          n_jobs=config["n_jobs"],
+                                          threshold=float(config["cluster_forming_alpha"]),
+                                          n_perm=config["n_permutations"])
+
+    np_outputs["logp_max_mass"].to_filename(np_logp_max_mass_path)
+
+    return np_logp_max_mass_path
