+++ connectomix/tests/tests_on_minimal_dataset/method-roiToVoxel.py
@@ -1,7 +1,7 @@
 import matplotlib.pyplot as plt
 import importlib.resources
 
-from connectomix.utils.modes import participant_level_analysis, group_level_analysis
+from connectomix.core.core import participant_level_pipeline, group_level_pipeline
 from connectomix.tests.paths import bids_dir, fmriprep_dir, output_dir
 
 precuneus_L_mask = str(importlib.resources.files("connectomix.tests.seeds").joinpath("AAL_Precuneus_L.nii.gz"))
@@ -12,7 +12,7 @@ roi_masks = {"precentralR": precentral_R_mask,
              "precuneusL": precuneus_L_mask}
 
 # Participant
-participant_level_analysis(bids_dir,
+participant_level_pipeline(bids_dir,
                            output_dir,
                            derivatives={"fmriprep": fmriprep_dir},
                            config={"method": method,
@@ -20,7 +20,7 @@ participant_level_analysis(bids_dir,
 
 # Group
 ## One-sample t-test
-group_level_analysis(bids_dir,
+group_level_pipeline(bids_dir,
                      output_dir,
                      config={"method": method,
                              "roi_masks": roi_masks,
@@ -28,7 +28,7 @@ group_level_analysis(bids_dir,
                              "contrast": "intercept"})
 
 ## Two-sample t-test, unpaired
-group_level_analysis(bids_dir,
+group_level_pipeline(bids_dir,
                      output_dir,
                      config={"method": method,
                              "roi_masks": roi_masks,
@@ -38,7 +38,7 @@ group_level_analysis(bids_dir,
                              "add_intercept": False})
 
 ## Regression
-group_level_analysis(bids_dir,
+group_level_pipeline(bids_dir,
                      output_dir,
                      config={"method": method,
                              "roi_masks": roi_masks,
