+++ connectomix/tests/tests_on_minimal_dataset/method-seedToVoxel.py
@@ -1,7 +1,7 @@
 import matplotlib.pyplot as plt
 import importlib.resources
 
-from connectomix.utils.modes import participant_level_analysis, group_level_analysis
+from connectomix.core.core import participant_level_pipeline, group_level_pipeline
 from connectomix.tests.paths import bids_dir, fmriprep_dir, output_dir
 
 example_seeds_for_seedToVoxel = str(importlib.resources.files("connectomix.tests.seeds").joinpath("example_seeds_for_seedToVoxel.tsv"))
@@ -9,7 +9,7 @@ example_seeds_for_seedToVoxel = str(importlib.resources.files("connectomix.tests
 method = "seedToVoxel"
 
 # Participant
-participant_level_analysis(bids_dir,
+participant_level_pipeline(bids_dir,
                            output_dir,
                            derivatives={"fmriprep": fmriprep_dir},
                            config={"method": method,
@@ -17,7 +17,7 @@ participant_level_analysis(bids_dir,
 
 # Group
 ## One-sample t-test
-group_level_analysis(bids_dir,
+group_level_pipeline(bids_dir,
                      output_dir,
                      config={"method": method,
                              "seeds_file": example_seeds_for_seedToVoxel,
@@ -25,7 +25,7 @@ group_level_analysis(bids_dir,
                              "contrast": "intercept"})
 
 ## Two-sample t-test, unpaired
-group_level_analysis(bids_dir,
+group_level_pipeline(bids_dir,
                      output_dir,
                      config={"method": method,
                              "seeds_file": example_seeds_for_seedToVoxel,
@@ -35,7 +35,7 @@ group_level_analysis(bids_dir,
                              "add_intercept": False})
 
 ## Regression
-group_level_analysis(bids_dir,
+group_level_pipeline(bids_dir,
                      output_dir,
                      config={"method": method,
                              "seeds_file": example_seeds_for_seedToVoxel,
